Refactor where main weather storage is done
Previously the weather storage was done globally, as you
can imagine this wasn't probably a good idea.

Changed the storage to be done in the main function, pass
the weather storage to the parts that need it.

Removed header file definitions that weren't needed now

Improved print to screen formatting