Refactoring of logger functionality
Changed the way the logger worked, as to pass through to it
the type of log we want printed, and have the logger decide
if it prints it based on what config options are set
from the command line.
Enums are Or'd into a variable and then checked in the logger
to see if they are enabled.